Split tests into test submodule.
